feat(ui): dashboard redesign round 1 — the answer strip
Merge HeroStripCard + the Rate-limit pace card into one
PaceOverviewStrip: each pace tile now contains the chart that explains
its number (hero used%/pace%, status + burn chips, the pace chart, and
the engine's outlook caption) instead of printing the same numbers
twice across two card rows ~400pt apart.

- Whole pace tile is clickable → compare-models modal (same affordance
  grammar as the Today tile → day detail); share stays hover-revealed.
- "via oauth · just now" freshness chip moves to the page header next
  to the advisor badges — it describes the page, not one card.
- Frees roughly a full card of fold space: Today-by-hour is now
  visible without scrolling at the default window size.
- Perf patterns preserved: synchronous bucketing per body pass,
  scan-tick @State caches, engine refresh only on recompute notify.

Awaiting Eric's sign-off on the before/after render before merge.
