--- App/Views/HeroStripCard.swift
@@ -1,567 +1,0 @@
-import SwiftUI
-import SwiftData
-import PacerCore
-import PacerUI
-
-/// "Show me everything important without scrolling" — the topmost card
-/// on the Dashboard. Three side-by-side tiles that answer the three
-/// questions the user opens Pacer to ask:
-///
-///   1. How much have I spent today?
-///   2. Am I going to hit the 5-hour limit?
-///   3. Am I going to hit the 7-day limit?
-///
-/// Replaces the prior dashboard pattern of stacking TodaySummaryCard +
-/// the gauges card on top of each other. Surfaces the same numbers with
-/// stronger visual hierarchy: the hero number per tile, the "/ pace"
-/// secondary number, a small chip that says behind/on/ahead/danger.
-struct HeroStripCard: View {
-    /// Optional click handler for the cost (Today) tile. When set, the
-    /// tile reads as actionable (cursor + hover) and forwards to the
-    /// caller — the dashboard wires this to opening today's day-detail
-    /// modal so the hero strip doubles as the primary "drill into
-    /// today" entry point.
-    let onTodayTap: (() -> Void)?
-
-    /// User's daily-cost notification toggle. Surfaced inside the
-    /// cost tile as a budget progress bar when enabled — the user
-    /// has opted into "I care about this number," so showing them
-    /// visual feedback (not just an after-the-fact banner) is
-    /// honest. When disabled, no bar — keeps the tile clean for
-    /// users who haven't configured a budget.
-    @AppStorage(PacerSettings.Key.notifyOnDailyCost, store: PacerSettings.store)
-    private var dailyBudgetEnabled: Bool = false
-    @AppStorage(PacerSettings.Key.dailyCostThresholdUSD, store: PacerSettings.store)
-    private var dailyBudgetUSD: Double = 50
-
-    init(onTodayTap: (() -> Void)? = nil) {
-        self.onTodayTap = onTodayTap
-        let todayString = TokenSample.formatDate(Date())
-        _todayAggregates = Query(
-            filter: #Predicate<DailyAggregate> { $0.date == todayString }
-        )
-    }
-
-    /// All today's per-model rollups. ≤ a handful of rows; iterating in
-    /// body is sub-millisecond.
-    @Query private var todayAggregates: [DailyAggregate]
-    /// Recent rate-limit samples — used only for the *current state*
-    /// snapshot per tile (used %, reset time). The burn projection itself
-    /// now comes from the intelligence engine, which owns the history.
-    @Query(HeroStripCard.recentRateLimits) private var rateLimits: [RateLimitSample]
-    /// The shared intelligence engine — source of the burn outlook (rate +
-    /// projected limit-hit) shown in each pace tile. Same answer the burn
-    /// notification uses, so tile and notification can't disagree.
-    @Environment(\.usageEngine) private var engine
-    /// Most-recent `extra_usage` sample. Only meaningful for Max-plan
-    /// users who can exceed quota; nil for everyone else. fetchLimit
-    /// = 1 keeps the materialization bounded.
-    @Query(HeroStripCard.recentExtraUsage) private var extraUsages: [ExtraUsageSample]
-    /// Scan-meta probe — used as the "data changed" tick that triggers
-    /// the @State cache refresh. Per AGENTS.md performance invariants,
-    /// every derived value heavier than O(N=10) lives in @State and is
-    /// only recomputed when this value flips, NOT on every body pass.
-    @Query(HeroStripCard.scanMetaProbe) private var scanMeta: [ClaudeCodeMeta]
-
-    private static let recentRateLimits: FetchDescriptor<RateLimitSample> = {
-        var d = FetchDescriptor<RateLimitSample>(
-            sortBy: [SortDescriptor(\.sampledAt, order: .reverse)]
-        )
-        d.fetchLimit = 48
-        return d
-    }()
-
-    private static let recentExtraUsage: FetchDescriptor<ExtraUsageSample> = {
-        var d = FetchDescriptor<ExtraUsageSample>(
-            sortBy: [SortDescriptor(\.sampledAt, order: .reverse)]
-        )
-        d.fetchLimit = 1
-        return d
-    }()
-
-    private static let scanMetaProbe: FetchDescriptor<ClaudeCodeMeta> = {
-        let key = ClaudeCodeMetaKey.lastIncrementalScanAt
-        return FetchDescriptor<ClaudeCodeMeta>(
-            predicate: #Predicate<ClaudeCodeMeta> { $0.key == key }
-        )
-    }()
-
-    // MARK: - @State cache (refreshed on scan-meta tick)
-
-    /// Snapshot of every derived value the body reads. Refreshed in
-    /// `refreshCache()` from the @Query data; the body itself never
-    /// iterates rate-limit rows or week aggregates.
-    @State private var cached = Cached()
-
-    private struct Cached {
-        var todayCost: Double = 0
-        var todayTokens: Int64 = 0
-        var fiveHour: SampleSnapshot?
-        var sevenDay: SampleSnapshot?
-        /// Latest `extra_usage` value in USD. nil when Anthropic
-        /// omitted the field; zero when we received it but it's at
-        /// zero (no overage yet). Both render as "no chip"; we
-        /// distinguish in the code path so we know whether to log.
-        var extraUsageUSD: Double?
-    }
-
-    /// Engine-produced answers, refreshed when the engine posts a fresh fit
-    /// (`.pacerEngineDidRecompute`) — not on raw scan ticks, so the tiles
-    /// never show an answer the engine hasn't computed yet.
-    @State private var fiveHourBurn: UsageIntelligenceEngine.BurnOutlook?
-    @State private var sevenDayBurn: UsageIntelligenceEngine.BurnOutlook?
-    /// Projected end-of-window utilization per window — what the burn chip
-    /// says when no pre-reset limit hit is projected ("≈52% at reset").
-    @State private var fiveHourEnd: Estimate?
-    @State private var sevenDayEnd: Estimate?
-    /// End-of-day projection + pace percentile for the Today tile's outlook
-    /// line ("≈$650 by tonight" once actionable; pace-vs-normal before that).
-    @State private var todayEOD: Estimate?
-    @State private var pacePercentile: Double?
-    /// Spend-so-far ÷ typical-by-this-hour for today's day-type — the
-    /// engine's most insightful midday number.
-    @State private var paceVsNow: Double?
-    /// Pace-ladder hysteresis (the label can't flap intra-day).
-    @State private var heldLadder: Int?
-
-    /// Sendable extract of the rate-limit row's display-relevant fields.
-    /// SwiftData @Model classes aren't Sendable, so we copy what we
-    /// need into a plain struct at refresh time.
-    private struct SampleSnapshot {
-        let window: String
-        let usedPercentage: Double
-        let resetsAt: Date?
-    }
-
-    private func refreshCache() {
-        var next = Cached()
-        next.todayCost = todayAggregates.reduce(0) { $0 + $1.totalCostUSD }
-        next.todayTokens = todayAggregates.reduce(0) {
-            $0 + $1.inputTokens + $1.outputTokens
-        }
-        if let s = rateLimits.first(where: { $0.window == "five_hour" }) {
-            next.fiveHour = SampleSnapshot(
-                window: s.window,
-                usedPercentage: s.usedPercentage,
-                resetsAt: s.resetsAt
-            )
-        }
-        if let s = rateLimits.first(where: { $0.window == "seven_day" }) {
-            next.sevenDay = SampleSnapshot(
-                window: s.window,
-                usedPercentage: s.usedPercentage,
-                resetsAt: s.resetsAt
-            )
-        }
-        if let latest = extraUsages.first {
-            next.extraUsageUSD = latest.amountUSD
-        }
-        cached = next
-    }
-
-    /// Re-ask the engine for both windows' burn outlooks + today's outlook.
-    private func refreshBurn() async {
-        guard let engine else { return }
-        fiveHourBurn = await engine.burnOutlook(window: .fiveHour)
-        sevenDayBurn = await engine.burnOutlook(window: .sevenDay)
-        fiveHourEnd = await engine.ask(.rateLimitOutlook(.fiveHour))
-        sevenDayEnd = await engine.ask(.rateLimitOutlook(.sevenDay))
-        todayEOD = await engine.ask(.projectedCost(.today))
-        let vsNow = await engine.ask(.paceVsNow)
-        paceVsNow = vsNow.isInsufficient ? nil : vsNow.value
-        let pace = await engine.ask(.pace)
-        pacePercentile = pace.isInsufficient ? nil : pace.value
-        if let p = pacePercentile {
-            heldLadder = IntelligenceFormatting.heldIndex(p, held: heldLadder)
-        }
-    }
-
-    /// The Today tile's one-line outlook chip — the single most insightful
-    /// statistic available right now, rendered as a colored chip:
-    ///   evening → "≈$650 by tonight" (anchors in the tooltip);
-    ///   midday, notably off-pace → "1.8× usual Friday spend" (arrow + semantic color);
-    ///   midday, on pace → "on your usual Friday pace" (.secondary);
-    ///   otherwise → the percentile ladder phrase (.secondary).
-    private var todayOutlook: (text: String, icon: String, tint: Color, help: String)? {
-        let dayName = Date().formatted(.dateTime.weekday(.wide))
-        if let e = todayEOD, IntelligenceFormatting.rangeIsActionable(e, spendSoFar: cached.todayCost) {
-            let help = e.interval80.map { IntelligenceFormatting.anchors($0) } ?? pacerCostExact(e.value)
-            return ("≈\(IntelligenceFormatting.approxCost(e.value)) by tonight",
-                    "chart.line.uptrend.xyaxis", .secondary, help)
-        }
-        if let r = paceVsNow {
-            if r >= 1.3 || (r <= 0.7 && cached.todayCost > 1) {
-                let help = "Dollars spent so far today vs the median spent by this hour on your \(dayName)s"
-                let icon: String
-                let tint: Color
-                if r >= 1.3 {
-                    icon = "arrow.up.right"
-                    tint = r >= 3 ? .red : .orange
-                } else {
-                    icon = "arrow.down.right"
-                    tint = .green
-                }
-                return ("\(IntelligenceFormatting.multiple(r)) usual \(dayName) spend",
-                        icon, tint, help)
-            }
-            if r > 0 {
-                let help = "Spend so far vs the median spend by this hour on your \(dayName)s"
-                return ("on your usual \(dayName) pace", "equal", .secondary, help)
-            }
-        }
-        if let p = pacePercentile {
-            let label = IntelligenceFormatting.paceLabel(
-                index: IntelligenceFormatting.heldIndex(p, held: heldLadder),
-                dayName: dayName)
-            return (label, "waveform.path.ecg", .secondary,
-                    "Today's projected total sits at the \(Int((p * 100).rounded()))th percentile of your days")
-        }
-        return nil
-    }
-
-    var body: some View {
-        HStack(spacing: 12) {
-            costTile
-            paceTile(
-                label: "5-hour pace",
-                sample: cached.fiveHour,
-                burn: fiveHourBurn,
-                endEstimate: fiveHourEnd,
-                duration: 5 * 3600
-            )
-            paceTile(
-                label: "7-day pace",
-                sample: cached.sevenDay,
-                burn: sevenDayBurn,
-                endEstimate: sevenDayEnd,
-                duration: 7 * 86400
-            )
-        }
-        .onAppear { refreshCache() }
-        .onChange(of: scanMeta.first?.value) { _, _ in refreshCache() }
-        // OAuth poller writes RateLimitSample without bumping scanMeta
-        // (only the JSONL scanner does that), so without this the hero
-        // tiles stay stale across successful OAuth polls. Watching the
-        // newest sample's sampledAt fires the refresh exactly when a
-        // new poll lands.
-        .onChange(of: rateLimits.first?.sampledAt) { _, _ in refreshCache() }
-        .task { await refreshBurn() }
-        .onReceive(NotificationCenter.default.publisher(for: .pacerEngineDidRecompute)) { _ in
-            Task { await refreshBurn() }
-        }
-    }
-
-    // MARK: - Cost tile
-
-    private var costTile: some View {
-        HeroTile(label: "Today", onTap: onTodayTap) {
-            VStack(alignment: .leading, spacing: 8) {
-                Text(pacerCost(cached.todayCost)).help(pacerCostExact(cached.todayCost))
-                    .font(.system(size: 32, weight: .semibold, design: .rounded))
-                    .monospacedDigit()
-                    .lineLimit(1)
-                    .minimumScaleFactor(0.7)
-                // Chip carries the outlook with semantic color + direction
-                // arrow so the user gets at-a-glance meaning without
-                // reading the words — same data as the old plain text
-                // line, but scannable.
-                if let o = todayOutlook {
-                    Chip(text: o.text, systemImage: o.icon, tint: o.tint, size: .compact)
-                        .fixedSize()
-                        .help(o.help)
-                }
-                HStack(spacing: 6) {
-                    if let extra = cached.extraUsageUSD, extra > 0 {
-                        // Max-plan overage. Tint orange (significant but
-                        // not catastrophic — at Anthropic's metered
-                        // rates a few dollars over isn't an emergency).
-                        Chip(
-                            text: "+\(pacerCost(extra)) over",
-                            systemImage: "plus.forwardslash.minus",
-                            tint: .orange,
-                            size: .compact
-                        )
-                        .fixedSize()
-                        .help("Over the included quota by \(pacerCostExact(extra))")
-                    }
-                    Text("\(pacerTokens(cached.todayTokens)) tokens")
-                        .font(.system(size: 11, weight: .medium))
-                        .foregroundStyle(.secondary)
-                        .lineLimit(1)
-                        .help("\(pacerTokensExact(cached.todayTokens)) tokens")
-                }
-                if dailyBudgetEnabled, dailyBudgetUSD > 0 {
-                    budgetBar
-                }
-            }
-        }
-    }
-
-    /// Progress bar showing today's cost as a fraction of the user's
-    /// configured daily budget. Tints green (<80%), orange (≥80%),
-    /// red (≥100%). The label says "$12 / $50 (24%)" — both raw and
-    /// percentage so the user doesn't have to mental-math either
-    /// side from the other.
-    @ViewBuilder
-    private var budgetBar: some View {
-        let ratio = dailyBudgetUSD > 0 ? cached.todayCost / dailyBudgetUSD : 0
-        let displayRatio = min(1.0, max(0, ratio))
-        let pctText = "\(Int((ratio * 100).rounded()))%"
-        let tint: Color =
-            ratio >= 1.0 ? .red :
-            ratio >= 0.8 ? .orange : .green
-        VStack(alignment: .leading, spacing: 3) {
-            HStack(spacing: 4) {
-                Text("\(pacerCost(cached.todayCost)) / \(pacerCost(dailyBudgetUSD))")
-                    .font(.system(size: 10, weight: .medium))
-                    .foregroundStyle(.secondary)
-                    .monospacedDigit()
-                    .help("\(pacerCostExact(cached.todayCost)) of \(pacerCostExact(dailyBudgetUSD)) daily budget")
-                Spacer(minLength: 0)
-                Text(pctText)
-                    .font(.system(size: 10, weight: .semibold, design: .rounded))
-                    .monospacedDigit()
-                    .foregroundStyle(tint)
-            }
-            GeometryReader { geo in
-                ZStack(alignment: .leading) {
-                    RoundedRectangle(cornerRadius: 2)
-                        .fill(Color.primary.opacity(0.08))
-                        .frame(height: 4)
-                    RoundedRectangle(cornerRadius: 2)
-                        .fill(tint)
-                        .frame(width: geo.size.width * displayRatio, height: 4)
-                }
-            }
-            .frame(height: 4)
-        }
-    }
-
-    // MARK: - Pace tile
-
-    @ViewBuilder
-    private func paceTile(
-        label: String,
-        sample: SampleSnapshot?,
-        burn: UsageIntelligenceEngine.BurnOutlook?,
-        endEstimate: Estimate?,
-        duration: TimeInterval
-    ) -> some View {
-        HeroTile(label: label) {
-            if let s = sample, let resets = s.resetsAt {
-                let cycle = DisplayCycle.resolve(resetsAt: resets, duration: duration)
-                if cycle.isAwaiting {
-                    // Sample's cycle already ended; no fresh sample yet.
-                    // Show "awaiting" rather than rendering stale used/pace %.
-                    awaitingTile
-                } else {
-                    let pacePct = cycle.paceFraction * 100
-                    let band = PaceBand(usedPct: s.usedPercentage, paceEndPct: pacePct)
-                    // Chip moved to its own row below the % line so it
-                    // can't ever wrap mid-word ("behi nd") when the tile
-                    // gets narrow. Layout: hero %, status row, optional
-                    // burn-rate forecast row.
-                    VStack(alignment: .leading, spacing: 6) {
-                        HStack(alignment: .firstTextBaseline, spacing: 4) {
-                            Text("\(Int(s.usedPercentage.rounded()))%")
-                                .font(.system(size: 32, weight: .semibold, design: .rounded))
-                                .monospacedDigit()
-                                .foregroundStyle(color(for: band))
-                                .lineLimit(1)
-                            Text("/")
-                                .font(.system(size: 18))
-                                .foregroundStyle(.tertiary)
-                            Text("\(Int(pacePct.rounded()))%")
-                                .font(.system(size: 18, weight: .medium, design: .rounded))
-                                .monospacedDigit()
-                                .foregroundStyle(.secondary)
-                                .lineLimit(1)
-                        }
-                        HStack(spacing: 8) {
-                            paceChip(band: band)
-                            Text("resets \(pacerRelative(resets))")
-                                .font(.system(size: 11, weight: .medium))
-                                .foregroundStyle(.secondary)
-                                .lineLimit(1)
-                        }
-                        burnRow(burn, endEstimate: endEstimate, duration: duration)
-                    }
-                }
-            } else {
-                VStack(alignment: .leading, spacing: 8) {
-                    Text("—")
-                        .font(.system(size: 32, weight: .semibold, design: .rounded))
-                        .foregroundStyle(.tertiary)
-                    Text("collecting…")
-                        .font(.system(size: 11, weight: .medium))
-                        .foregroundStyle(.secondary)
-                }
-            }
-        }
-    }
-
-    /// Layout matches the active-cycle tile's vertical footprint so the
-    /// HStack's equal-height layout doesn't reflow when a sibling tile
-    /// is showing active data.
-    private var awaitingTile: some View {
-        VStack(alignment: .leading, spacing: 6) {
-            Text("—")
-                .font(.system(size: 32, weight: .semibold, design: .rounded))
-                .foregroundStyle(.tertiary)
-                .lineLimit(1)
-            Text("awaiting first sample of new cycle")
-                .font(.system(size: 11, weight: .medium))
-                .foregroundStyle(.secondary)
-                .lineLimit(2)
-            Text(" ")  // matches burnRow's empty-placeholder height
-                .font(.system(size: 11))
-                .opacity(0)
-        }
-    }
-
-    /// Optional burn-outlook row inside a pace tile, rendered as a single
-    /// chip: semantic color (green/secondary/orange/red) + flame icon, and
-    /// text that says what *happens* rather than a rate ratio — "limit in
-    /// 7 hr" when a pre-reset hit is projected, "≈52% at reset" otherwise.
-    /// (The old "0.5× sustainable" framing read as jargon.) Raw %/hr numbers
-    /// live in the tooltip.
-    @ViewBuilder
-    private func burnRow(
-        _ projection: UsageIntelligenceEngine.BurnOutlook?,
-        endEstimate: Estimate?,
-        duration: TimeInterval
-    ) -> some View {
-        if let projection,
-           let chip = Self.burnChip(projection: projection, endEstimate: endEstimate, duration: duration) {
-            Chip(text: chip.text, systemImage: "flame.fill", tint: chip.tint, size: .compact)
-                .fixedSize()
-                .help(chip.help)
-        } else {
-            // Empty placeholder of the same vertical footprint so the
-            // tile height matches its sibling whose burn rate did
-            // surface — otherwise the HStack's equal-height layout
-            // would stretch the shorter tile and look misaligned.
-            Text(" ")
-                .font(.system(size: 11))
-                .opacity(0)
-        }
-    }
-
-    /// Chip text + tint + tooltip for a burn outlook. nil when effectively
-    /// idle (chip hidden). Plain function because branch-assignments don't
-    /// belong in a ViewBuilder.
-    private static func burnChip(
-        projection: UsageIntelligenceEngine.BurnOutlook,
-        endEstimate: Estimate?,
-        duration: TimeInterval
-    ) -> (text: String, tint: Color, help: String)? {
-        let ratio = IntelligenceFormatting.capPaceRatio(
-            slopePercentPerHour: projection.slopePercentPerHour, windowSeconds: duration)
-        guard ratio >= 0.05 else { return nil }          // effectively idle — say nothing
-        var help = IntelligenceFormatting.capPaceHelp(
-            slopePercentPerHour: projection.slopePercentPerHour, windowSeconds: duration)
-        if projection.willHitLimitBeforeReset,
-           let eta = IntelligenceFormatting.relativeCrossingPhrase(projection) {
-            return ("limit \(eta)", .red, help)
-        }
-        let tint: Color = ratio < 0.9 ? .green : ratio < 1.15 ? .secondary : .orange
-        if let e = endEstimate, !e.isInsufficient {
-            if let band = e.interval80 {
-                help += String(format: " Calibrated range at reset: %.0f–%.0f%%.",
-                               band.lowerBound, band.upperBound)
-            }
-            return ("≈\(Int(e.value.rounded()))% at reset", tint, help)
-        }
-        // No calibrated end estimate yet (young history): fall back to the
-        // raw rate, which at least is a real unit.
-        return (String(format: "%+.1f%%/hr", projection.slopePercentPerHour), tint, help)
-    }
-
-    @ViewBuilder
-    private func paceChip(band: PaceBand) -> some View {
-        Group {
-            switch band {
-            case .green:
-                Chip(text: "behind", systemImage: "checkmark", tint: .green, size: .compact)
-            case .white:
-                Chip(text: "on pace", tint: .secondary, size: .compact)
-            case .yellow:
-                Chip(text: "ahead", systemImage: "exclamationmark", tint: .yellow, size: .compact)
-            case .red:
-                Chip(text: "danger", systemImage: "exclamationmark.triangle.fill", tint: .red, size: .compact)
-            }
-        }
-        // Lock the chip's width to its content so it can't get squeezed
-        // and wrap. Layout above uses a Spacer/HStack that would
-        // otherwise let the chip's text wrap when the tile narrows.
-        .fixedSize()
-    }
-
-    private func color(for band: PaceBand) -> Color {
-        switch band {
-        case .green:  return .green
-        case .white:  return .primary
-        case .yellow: return .yellow
-        case .red:    return .red
-        }
-    }
-}
-
-/// One tile in the hero strip. Cards-of-cards: each tile reuses the
-/// app's standard PacerCard surface so spacing/radius/stroke stay
-/// consistent with the rest of the dashboard.
-///
-/// When `onTap` is non-nil the tile becomes a hover-reactive button
-/// — same surface, but with a subtle hover ring + pointing-hand
-/// cursor so it reads as actionable. The tile remains a passive
-/// surface when no callback is wired.
-private struct HeroTile<Content: View>: View {
-    let label: String
-    var onTap: (() -> Void)? = nil
-    @ViewBuilder let content: () -> Content
-    @State private var hovering = false
-
-    var body: some View {
-        let surface = VStack(alignment: .leading, spacing: 10) {
-            HStack(spacing: 6) {
-                Eyebrow(text: label)
-                Spacer()
-                if onTap != nil {
-                    Image(systemName: "arrow.up.right.circle.fill")
-                        .font(.system(size: 12))
-                        .foregroundStyle(.tertiary)
-                        .opacity(hovering ? 1.0 : 0.0)
-                }
-            }
-            content()
-            Spacer(minLength: 0)
-        }
-        .padding(PacerDesign.cardPadding)
-        .frame(maxWidth: .infinity, minHeight: 116, alignment: .topLeading)
-        .background(
-            RoundedRectangle(cornerRadius: PacerDesign.cardCornerRadius, style: .continuous)
-                .fill(Color(nsColor: .controlBackgroundColor))
-        )
-        .overlay(
-            RoundedRectangle(cornerRadius: PacerDesign.cardCornerRadius, style: .continuous)
-                .stroke(
-                    onTap != nil && hovering
-                        ? Color.accentColor.opacity(0.45)
-                        : PacerDesign.cardStroke,
-                    lineWidth: 1
-                )
-        )
-
-        if let onTap {
-            Button(action: onTap) {
-                surface.contentShape(Rectangle())
-            }
-            .buttonStyle(.plain)
-            .onHover { hovering = $0 }
-            .help("Open today's breakdown")
-            .accessibilityHint("Opens today's detailed breakdown")
-        } else {
-            surface
-        }
-    }
-}
